# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: MTOGO API Endpoints
  description: API through the gateway for managing restaurants, customers, and management operations on the MTOGO platform. Made by Andreas and Owais.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
...
paths:
...
  /api/basket:
    get:
      tags:
        - Customers
      summary: Get the customer's basket
      description: Retrieves the contents of a customer's basket for a specific restaurant.
      parameters:
        - name: restaurantId
          in: query
          required: true
          description: The ID of the restaurant for which to retrieve the basket.
          schema:
            type: string
            example: cm49qi0zr00006122j1hml4gm
      responses:
        '200':
          description: Basket retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  basket:
                    type: array
                    items:
                      type: object
                      properties:
                        itemId:
                          type: string
                          example: 'item12345'
                        name:
                          type: string
                          example: 'Pizza Margherita'
                        quantity:
                          type: integer
                          example: 2
                        price:
                          type: number
                          format: float
                          example: 12.99
        '404':
          description: Basket not found for the given customer and restaurant.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Basket not found for the given customer and restaurant.'
        '401':
          description: Unauthorized request. Session cookie required.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized. Session cookie required.'
        '500':
          description: Internal server error.

    post:
      tags:
        - Customers
      summary: Add an item to the customer's basket
      description: Allows a customer to add a menu item to their basket for a specific restaurant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                menuId:
                  type: string
                  example: cm49qjxir00041zrz5hbephtu
                  description: The ID of the menu item to add to the basket.
                title:
                  type: string
                  example: Vegetarian
                  description: The title/name of the menu item.
                quantity:
                  type: integer
                  example: 2
                  description: The number of items to add to the basket.
                restaurantId:
                  type: string
                  example: cm49qi0zr00006122j1hml4gm
                  description: The ID of the restaurant associated with the basket.
      responses:
        '200':
          description: Item added successfully to the basket.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Item added to the basket successfully.'
                  basket:
                    type: array
                    items:
                      type: object
                      properties:
                        itemId:
                          type: string
                          example: 'item12345'
                        name:
                          type: string
                          example: 'Pizza Margherita'
                        quantity:
                          type: integer
                          example: 2
                        price:
                          type: number
                          format: float
                          example: 12.99
        '400':
          description: Invalid input data or missing fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Invalid input data. Please provide all required fields.'
        '404':
          description: Restaurant or menu not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Restaurant or menu not found.'
        '401':
          description: Unauthorized request. Session cookie required.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized. Session cookie required.'
        '500':
          description: Internal server error.

    put:
      tags:
        - Customers
      summary: Update an item in the customer's basket
      description: Allows a customer to update the quantity of a menu item in their basket for a specific restaurant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itemId:
                  type: string
                  example: cm3vqlg1t0006j48as8novj65
                  description: The ID of the item in the basket to update.
                quantity:
                  type: integer
                  example: 0
                  description: The updated quantity of the item in the basket. A value of 0 may indicate removal from the basket.
                restaurantId:
                  type: string
                  example: cm3vqicwm000134choat5sho7
                  description: The ID of the restaurant associated with the basket.
      responses:
        '200':
          description: Item updated successfully in the basket.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Item updated successfully in the basket.'
                  basket:
                    type: array
                    items:
                      type: object
                      properties:
                        itemId:
                          type: string
                          example: 'item12345'
                        name:
                          type: string
                          example: 'Pizza Margherita'
                        quantity:
                          type: integer
                          example: 2
                        price:
                          type: number
                          format: float
                          example: 12.99
        '400':
          description: Invalid input data or missing fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Invalid input data. Please provide all required fields.'
        '404':
          description: Item or restaurant not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Item or restaurant not found.'
        '401':
          description: Unauthorized request. Session cookie required.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized. Session cookie required.'
        '500':
          description: Internal server error.

    delete:
      tags:
        - Customers
      summary: Clear the customer's basket
      description: Allows a customer to remove all items from their basket for a specific restaurant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                restaurantId:
                  type: string
                  example: cm3vqicwm000134choat5sho7
                  description: The ID of the restaurant associated with the basket.
      responses:
        '200':
          description: Basket cleared successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Basket cleared successfully.'
        '404':
          description: Restaurant not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Restaurant not found.'
        '401':
          description: Unauthorized request. Session cookie required.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized. Session cookie required.'
        '500':
          description: Internal server error.
  /api/orders:
    post:
      tags:
        - Customers
      summary: Create an order
      description: Allows a customer to create an order using an existing basket and specify delivery and payment details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                basketId:
                  type: string
                  example: cm4a3xyjb000813rxpakc76yk
                  description: The ID of the basket to use for the order.
                deliveryAddress:
                  type: object
                  properties:
                    recipentName:
                      type: string
                      example: Andreas Fritzbøger
                      description: The name of the person receiving the delivery.
                    street:
                      type: string
                      example: Apple Blvardi
                      description: The street address for delivery.
                    city:
                      type: string
                      example: New Yorky
                      description: The city for the delivery address.
                    zip:
                      type: string
                      example: '2380'
                      description: The postal code for the delivery address.
                    floor:
                      type: string
                      example: 5th
                      description: The floor or additional details for the delivery address.
                payment:
                  type: object
                  properties:
                    method:
                      type: string
                      enum:
                        - VISA
                        - MASTER_CARD
                      example: VISA
                      description: The payment method for the order.
      responses:
        '200':
          description: Order created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Order created successfully
                  order:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 87d06907-29b2-465e-bd65-78b08a0a0333
                      customerId:
                        type: string
                        example: cm49qj3gh00026122dz8kefie
                      restaurantId:
                        type: string
                        example: cm49qi0zr00006122j1hml4gm
                      status:
                        type: string
                        example: YOUR_FOOD_IS_BEING_PREPARED
                      totalAmount:
                        type: number
                        format: float
                        example: 258
                      paymentIntentId:
                        type: string
                        example: pi_3QSLWrDAx3Y1tFqk1lQrGtnP
                      note:
                        type: string
                        example: null
                      createdAt:
                        type: string
                        format: date-time
                        example: '2024-12-04T16:34:22.447Z'
                      updatedAt:
                        type: string
                        format: date-time
                        example: '2024-12-04T16:34:22.447Z'
                      deliveryAddressesId:
                        type: string
                        example: 8e86cf43-d54b-482b-a1b2-dd94702de314
        '400':
          description: Invalid input data or missing fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid input data. Please provide all required fields.
        '404':
          description: Basket not found or invalid restaurant or payment method.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Basket not found or invalid restaurant/payment method.
        '401':
          description: Unauthorized request. Session cookie required.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized. Session cookie required.
        '500':
          description: Internal server error.
...